+++ kxs/src/commonMain/kotlin/at/asitplus/awesn1/serialization/internal/DerEncoder.kt
@@ -61,7 +61,7 @@ class DerEncoder internal constructor(
 
     @OptIn(ExperimentalSerializationApi::class)
     override fun encodeInline(descriptor: SerialDescriptor): Encoder {
-        inlineHintState.recordFrom(descriptor)
+        inlineHintState.captureInlineHintsFrom(descriptor)
         return this
     }
 
@@ -227,6 +227,10 @@ class DerEncoder internal constructor(
      */
     @Throws(SerializationException::class)
     override fun <T> encodeSerializableValue(serializer: SerializationStrategy<T>, value: T) {
+        if (value != null && serializer.descriptor.isInline) {
+            super<AbstractEncoder>.encodeSerializableValue(serializer, value)
+            return
+        }
 
         val inlineHints = inlineHintState.consume()
         val propertyContext = descriptorAndIndex?.toDerPropertyContext()